Give UTF-8 files a BOM man! Always!

+++ src/FlatfileExport.php
@@ -2,13 +2,13 @@
 
 namespace LaravelFlatfiles;
 
-use League\Csv\Writer;
-use Illuminate\Support\Arr;
-use Illuminate\Support\Str;
-use Illuminate\Support\Collection;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Support\Facades\Storage;
 use Illuminate\Filesystem\FilesystemAdapter;
+use Illuminate\Support\Arr;
+use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
+use League\Csv\Writer;
 
 class FlatfileExport
 {
@@ -86,7 +86,7 @@ class FlatfileExport
             throw new \RuntimeException('Target export file already exists at: '.$absoluteFilepath);
         }
 
-        if (! file_exists(dirname($absoluteFilepath))) {
+        if (!file_exists(dirname($absoluteFilepath))) {
             mkdir($absoluteFilepath, 0777, true);
         }
 
@@ -169,9 +169,17 @@ class FlatfileExport
     {
         $this->addBomIfNeeded();
 
+        if (!$this->usesDisk()) {
+            if ($this->pathToLocalTmpFile == $this->pathToFile()) {
+                return true;
+            }
+
+            return rename($this->pathToLocalTmpFile, $this->pathToFile());
+        }
+
         $this->disk()->putStream($this->pathToFile(), fopen($this->pathToLocalTmpFile, 'r'));
 
-        unlink($this->pathToLocalTmpFile);
+        return unlink($this->pathToLocalTmpFile);
     }
 
     /**
@@ -184,7 +192,7 @@ class FlatfileExport
 
         switch ($extension = $this->targetfileExtension()) {
             case 'csv':
-                if (! $this->pathToLocalTmpFile) {
+                if (!$this->pathToLocalTmpFile) {
                     if ($this->usesDisk()) {
                         $this->pathToLocalTmpFile = tempnam(sys_get_temp_dir(), 'ffe');
                     } else {
@@ -252,7 +260,7 @@ class FlatfileExport
      */
     private function makeModelAttributesVisible($model)
     {
-        if (! ($model instanceof Model)) {
+        if (!($model instanceof Model)) {
             return $model;
         }
 
@@ -266,7 +274,7 @@ class FlatfileExport
 
     private function addBomIfNeeded()
     {
-        if ($this->bomNeedsToBeAdded && ! $this->checkbom()) {
+        if ($this->bomNeedsToBeAdded && !$this->checkbom()) {
             file_put_contents($this->pathToLocalTmpFile, Writer::BOM_UTF8.file_get_contents($this->pathToLocalTmpFile));
             $this->bomNeedsToBeAdded = false;
         }
